% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist-matrix}
\alias{dist-matrix}
\alias{distR}
\title{Distance Matrix Computation from Two Matrices}
\usage{
distR(x1, x2 = NULL)
}
\arguments{
\item{x1}{matrix of size N1 x n}

\item{x2}{matrix of size N2 x n}
}
\value{
Matrix of size N1 x N2
}
\description{
This function extends \code{dist} to accept two arguments.
}
\details{
Computes the distances between the coordinates in \code{x1} and the coordinates in \code{x2}. The matrices \code{x1} and \code{x2} do not need to have the same number of rows, but need to have the same number of columns (dimensions).
}
\examples{
A <- matrix(rnorm(50),5,10)
D <- distR(A,A[-3,])
}
